+++ C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/service/post/IPostService.java
@@ -1,4 +1,5 @@
 package com.c0722g1repobe.service.post;
+import com.c0722g1repobe.dto.post.PostDto;
 import com.c0722g1repobe.dto.post.PostListViewDto;
 import com.c0722g1repobe.entity.post.Post;
 import org.springframework.data.domain.Page;
@@ -6,13 +7,7 @@ import org.springframework.data.domain.Pageable;
 
 public interface IPostService {
     Page<PostListViewDto> findAll(String area, String price, String demandType, String direction, String city, Pageable pageable);
-import com.c0722g1repobe.dto.post.PostDto;
-import com.c0722g1repobe.entity.customer.Customer;
-import org.springframework.data.repository.query.Param;
 
-import java.util.Optional;
-
-public interface IPostService {
     /**
      * Create by: NgocLV
      * Date created: 31/01/2023
@@ -53,7 +48,7 @@ public interface IPostService {
      *
      * @return json list posts
      */
-    Page<PostDto> findAllPost( Pageable pageable);
+    Page<PostDto> findAllPost(Pageable pageable);
     /**
      * Create by: NgocLV
      * Date created: 31/01/2023
